Create Lower to upper case.c
In this program, the user is asked to enter a string, which is stored in the str array. Then, a for loop is used to traverse through the string, and if a character is lowercase, it is converted to uppercase by subtracting 32 from its ASCII value. The converted string is displayed on the screen.