Create Check whether number is prime or not.c
In this program, the user is asked to enter a positive integer, which is stored in the num variable. Then, a for loop is used to check whether the number is divisible by any number between 2 and half of the number itself. If the number is divisible by any number within this range, then it is not a prime number. The flag variable is used to keep track of whether the number is divisible by any other number or not.