Create Printing abbreviation of a given name.c
This program first prompts the user to enter their full name, stores it in the name array, then uses a loop to iterate over the characters in the name. Whenever a space character is encountered, the next character is converted to uppercase using the toupper function and added to the abbreviation. Finally, the abbreviated name is printed.